Feature/farm

+++ src/main/java/com/cts/FoodChainX/dto/tracerecord/TraceRecordResponse.java
@@ -5,8 +5,8 @@ import java.time.LocalDate;
 
 @Builder
 public record TraceRecordResponse(
-    Integer traceId,
-    Integer batchId,
+    Long traceId,
+    Long batchId,
     String cropType,
     String farmName,
     String distributorName,
