+++ src/main/java/com/cts/FoodChainX/repository/TraceRecordRepository.java
@@ -1,12 +1,14 @@
 package com.cts.FoodChainX.repository;
 
-import com.cts.FoodChainX.model.TraceRecord;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.cts.FoodChainX.model.TraceRecord;
+
 @Repository
-public interface TraceRecordRepository extends JpaRepository<TraceRecord, Integer> {
+public interface TraceRecordRepository extends JpaRepository<TraceRecord, Long> {
     // Corrected to match ProductionBatch.productionId
-    Optional<TraceRecord> findByProductionBatch_ProductionId(Integer productionId);
+    Optional<TraceRecord> findByProductionBatch_ProductionId(Long productionId);
 }
